Enforce IWYU on things needed for core build
Client CLs:
 - http://ag/21408973
 - https://crrev.com/c/4247679
 - http://cl/509514434
 - http://cl/509515755

non-trivial changes to:
 - SkBitmap.h to allow forward declares of SkShader and SkTileMode.
 - SkCanvas.h to allow SkSpecialImage to be forward-declared.

Everything else was mechanical.

Bug: skia:13983
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/642658
Commit-Queue: Kevin Lubick <kjlubick@google.com>

+++ gm/runtimeshader.cpp
@@ -161,7 +161,7 @@ public:
     void onDraw(SkCanvas* canvas) override {
         SkRuntimeShaderBuilder builder(fEffect);
 
-        builder.uniform("cutoff") = sin(fSecs) * 0.55f + 0.5f;
+        builder.uniform("cutoff") = sinf(fSecs) * 0.55f + 0.5f;
         builder.uniform("slope")  = 10.0f;
 
         builder.child("before_map")    = fBefore;
